[of_stats] Refactoring
