+++ kytos/of_stats/__init__.py
@@ -1,0 +1,1 @@
+"""Statistics napp package."""
